Update index.html
This file is the root document of the website. On github, changing this file will change the URL https://co2pi.github.io/circularity/
However, the change is not fast, and how/when the change takes effect on the public URL is unclear. 
Let us assume that it takes 24 hours to update, who knows. 

Our project website is at http://entropynetwrork.com/circularity ; this updates immediately.